Fix assigned type and return value difference

+++ Dynamic Array/main.cpp
@@ -79,9 +79,9 @@ int main() {
 		arr[i] = input;
 	}
 
-	int max = computeMaximum(arr, size);
-	int min = computeMinimum(arr, size);
-	int avg = computeMinimum(arr, size);
+	float max = computeMaximum(arr, size);
+	float min = computeMinimum(arr, size);
+	float avg = computeMinimum(arr, size);
 
 	std::cout << "The maximum value of the array is: " << max << std::endl <<
 		"The minimum value of the array is: " << min << std::endl <<
